Moved COSE related things to their own package
