+++ protocol/attestation_u2f.go
@@ -6,6 +6,7 @@ import (
 	"crypto/elliptic"
 	"crypto/x509"
 
+	"github.com/duo-labs/webauthn/protocol/webauthncose"
 	"github.com/ugorji/go/codec"
 )
 
@@ -19,10 +20,10 @@ func init() {
 func verifyU2FFormat(att AttestationObject, clientDataHash []byte) (string, []interface{}, error) {
 	// Signing procedure step - If the credential public key of the given credential is not of
 	// algorithm -7 ("ES256"), stop and return an error.
-	key := EC2PublicKeyData{}
+	key := webauthncose.EC2PublicKeyData{}
 	codec.NewDecoder(bytes.NewReader(att.AuthData.AttData.CredentialPublicKey), new(codec.CborHandle)).Decode(&key)
 
-	if COSEAlgorithmIdentifier(key.PublicKeyData.Algorithm) != AlgES256 {
+	if webauthncose.COSEAlgorithmIdentifier(key.PublicKeyData.Algorithm) != webauthncose.AlgES256 {
 		return u2fAttestationKey, nil, ErrUnsupportedAlgorithm.WithDetails("Non-ES256 Public Key algorithm used")
 	}
