+++ protocol/authenticator.go
@@ -2,18 +2,10 @@ package protocol
 
 import (
 	"bytes"
-	"crypto"
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rsa"
-	"encoding/asn1"
 	"encoding/binary"
 	"fmt"
-	"hash"
-	"math/big"
 
 	"github.com/ugorji/go/codec"
-	"golang.org/x/crypto/ed25519"
 )
 
 var minAuthDataLength = 37
@@ -54,89 +46,6 @@ type AttestedCredentialData struct {
 	// The raw credential public key bytes received from the attestation data
 	CredentialPublicKey []byte `json:"public_key"`
 }
-
-// PublicKeyData The public key portion of a Relying Party-specific credential key pair, generated
-// by an authenticator and returned to a Relying Party at registration time. We unpack this object
-// using ugorji's codec library (github.com/ugorji/go/codec) which is why there are codec tags
-// included. The tag field values correspond to the IANA COSE keys that give their respective
-// values.
-// See §6.4.1.1 https://www.w3.org/TR/webauthn/#sctn-encoded-credPubKey-examples for examples of this
-// COSE data.
-type PublicKeyData struct {
-	// Decode the results to int by default.
-	_struct bool `codec:",int" json:"public_key"`
-	// The type of key created. Should be OKP, EC2, or RSA.
-	KeyType int64 `codec:"1" json:"kty"`
-	// A COSEAlgorithmIdentifier for the algorithm used to derive the key signature.
-	Algorithm int64 `codec:"3" json:"alg"`
-}
-type EC2PublicKeyData struct {
-	PublicKeyData
-	// If the key type is EC2, the curve on which we derive the signature from.
-	Curve int64 `codec:"-1,omitempty" json:"crv"`
-	// A byte string 32 bytes in length that holds the x coordinate of the key.
-	XCoord []byte `codec:"-2,omitempty" json:"x-coordinate"`
-	// A byte string 32 bytes in length that holds the y coordinate of the key.
-	YCoord []byte `codec:"-3,omitempty" json:"y-coordinate"`
-}
-
-type RSAPublicKeyData struct {
-	PublicKeyData
-	// Represents the modulus parameter for the RSA algorithm
-	Modulus []byte `codec:"-1,omitempty" json:"mod"`
-	// Represents the exponent parameter for the RSA algorithm
-	Exponent []byte `codec:"-2,omitempty" json:"exp"`
-}
-
-type OKPPublicKeyData struct {
-	PublicKeyData
-	Curve int64
-	// A byte string that holds the x coordinate of the key.
-	XCoord []byte `codec:"-2,omitempty" json:"x-coordinate"`
-}
-
-// COSEAlgorithmIdentifier From §5.10.5. A number identifying a cryptographic algorithm. The algorithm
-// identifiers SHOULD be values registered in the IANA COSE Algorithms registry
-// [https://www.w3.org/TR/webauthn/#biblio-iana-cose-algs-reg], for instance, -7 for "ES256"
-//  and -257 for "RS256".
-type COSEAlgorithmIdentifier int
-
-const (
-	// AlgES256 ECDSA with SHA-256
-	AlgES256 COSEAlgorithmIdentifier = -7
-	// AlgES384 ECDSA with SHA-384
-	AlgES384 COSEAlgorithmIdentifier = -35
-	// AlgES512 ECDSA with SHA-512
-	AlgES512 COSEAlgorithmIdentifier = -36
-	// AlgRS1 RSASSA-PKCS1-v1_5 with SHA-1
-	AlgRS1 COSEAlgorithmIdentifier = -65535
-	// AlgRS256 RSASSA-PKCS1-v1_5 with SHA-256
-	AlgRS256 COSEAlgorithmIdentifier = -257
-	// AlgRS384 RSASSA-PKCS1-v1_5 with SHA-384
-	AlgRS384 COSEAlgorithmIdentifier = -258
-	// AlgRS512 RSASSA-PKCS1-v1_5 with SHA-512
-	AlgRS512 COSEAlgorithmIdentifier = -259
-	// AlgPS256 RSASSA-PSS with SHA-256
-	AlgPS256 COSEAlgorithmIdentifier = -37
-	// AlgPS384 RSASSA-PSS with SHA-384
-	AlgPS384 COSEAlgorithmIdentifier = -38
-	// AlgPS512 RSASSA-PSS with SHA-512
-	AlgPS512 COSEAlgorithmIdentifier = -39
-	// AlgEdDSA EdDSA
-	AlgEdDSA COSEAlgorithmIdentifier = -8
-)
-
-// The Key Type derived from the IANA COSE AuthData
-type COSEKeyType int
-
-const (
-	// OctetKey is an Octet Key
-	OctetKey COSEKeyType = 1
-	// EllipticKey is an Elliptic Curve Public Key
-	EllipticKey COSEKeyType = 2
-	// RSAKey is an RSA Public Key
-	RSAKey COSEKeyType = 3
-)
 
 // AuthenticatorAttachment https://www.w3.org/TR/webauthn/#platform-attachment
 type AuthenticatorAttachment string
@@ -271,32 +180,6 @@ func (a *AuthenticatorData) Unmarshal(rawAuthData []byte) error {
 	return nil
 }
 
-// Figure out what kind of COSE material was provided and create the data for the new key
-func ParsePublicKey(keyBytes []byte) (interface{}, error) {
-	var cborHandler codec.Handle = new(codec.CborHandle)
-	pk := PublicKeyData{}
-	codec.NewDecoder(bytes.NewReader(keyBytes), cborHandler).Decode(&pk)
-	switch COSEKeyType(pk.KeyType) {
-	case OctetKey:
-		var o OKPPublicKeyData
-		codec.NewDecoder(bytes.NewReader(keyBytes), cborHandler).Decode(&o)
-		o.PublicKeyData = pk
-		return o, nil
-	case EllipticKey:
-		var e EC2PublicKeyData
-		codec.NewDecoder(bytes.NewReader(keyBytes), cborHandler).Decode(&e)
-		e.PublicKeyData = pk
-		return e, nil
-	case RSAKey:
-		var r RSAPublicKeyData
-		codec.NewDecoder(bytes.NewReader(keyBytes), cborHandler).Decode(&r)
-		r.PublicKeyData = pk
-		return r, nil
-	default:
-		return nil, ErrUnsupportedKey
-	}
-}
-
 // If Attestation Data is present, unmarshall that into the appropriate public key structure
 func (a *AuthenticatorData) unmarshalAttestedData(rawAuthData []byte) {
 	a.AttData.AAGUID = rawAuthData[37:53]
@@ -351,102 +234,4 @@ func (a *AuthenticatorData) Verify(rpIdHash []byte, userVerificationRequired boo
 	// This is not yet fully implemented by the spec or by browsers
 
 	return nil
-}
-
-// Verify Octet Key Pair (OKP) Public Key Signature
-func (k *OKPPublicKeyData) Verify(data []byte, sig []byte) (bool, error) {
-	f := HasherFromCOSEAlg(COSEAlgorithmIdentifier(k.PublicKeyData.Algorithm))
-	h := f()
-	h.Write(data)
-	return ed25519.Verify(k.XCoord, h.Sum(nil), sig), nil
-}
-
-// Verify Elliptic Curce Public Key Signature
-func (k *EC2PublicKeyData) Verify(data []byte, sig []byte) (bool, error) {
-	var curve elliptic.Curve
-	switch COSEAlgorithmIdentifier(k.Algorithm) {
-	case AlgES512: // IANA COSE code for ECDSA w/ SHA-512
-		curve = elliptic.P521()
-	case AlgES384: // IANA COSE code for ECDSA w/ SHA-384
-		curve = elliptic.P384()
-	case AlgES256: // IANA COSE code for ECDSA w/ SHA-256
-		curve = elliptic.P256()
-	default:
-		return false, ErrUnsupportedAlgorithm
-	}
-
-	pubkey := &ecdsa.PublicKey{
-		Curve: curve,
-		X:     big.NewInt(0).SetBytes(k.XCoord),
-		Y:     big.NewInt(0).SetBytes(k.YCoord),
-	}
-
-	type ECDSASignature struct {
-		R, S *big.Int
-	}
-
-	e := &ECDSASignature{}
-	f := HasherFromCOSEAlg(COSEAlgorithmIdentifier(k.PublicKeyData.Algorithm))
-	h := f()
-	h.Write(data)
-	_, error := asn1.Unmarshal(sig, e)
-	return ecdsa.Verify(pubkey, h.Sum(nil), e.R, e.S), error
-}
-
-// Verify RSA Public Key Signature
-func (k *RSAPublicKeyData) Verify(data []byte, sig []byte) (bool, error) {
-	pubkey := &rsa.PublicKey{
-		N: big.NewInt(0).SetBytes(k.Modulus),
-		E: int(uint(k.Exponent[2]) | uint(k.Exponent[1])<<8 | uint(k.Exponent[0])<<16),
-	}
-
-	f := HasherFromCOSEAlg(COSEAlgorithmIdentifier(k.PublicKeyData.Algorithm))
-	h := f()
-	h.Write(data)
-
-	var hash crypto.Hash
-	switch COSEAlgorithmIdentifier(k.PublicKeyData.Algorithm) {
-	case AlgRS1:
-		hash = crypto.SHA1
-	case AlgPS256, AlgRS256:
-		hash = crypto.SHA256
-	case AlgPS384, AlgRS384:
-		hash = crypto.SHA384
-	case AlgPS512, AlgRS512:
-		hash = crypto.SHA512
-	default:
-		return false, ErrUnsupportedAlgorithm
-	}
-	switch COSEAlgorithmIdentifier(k.PublicKeyData.Algorithm) {
-	case AlgPS256, AlgPS384, AlgPS512:
-		err := rsa.VerifyPSS(pubkey, hash, h.Sum(nil), sig, nil)
-		return err == nil, err
-
-	case AlgRS1, AlgRS256, AlgRS384, AlgRS512:
-		err := rsa.VerifyPKCS1v15(pubkey, hash, h.Sum(nil), sig)
-		return err == nil, err
-	default:
-		return false, ErrUnsupportedAlgorithm
-	}
-}
-
-// Return which signature algorithm is being used from the COSE Key
-func SigAlgFromCOSEAlg(coseAlg COSEAlgorithmIdentifier) SignatureAlgorithm {
-	for _, details := range signatureAlgorithmDetails {
-		if details.coseAlg == coseAlg {
-			return details.algo
-		}
-	}
-	return UnknownSignatureAlgorithm
-}
-
-// Return the Hashing interface to be used for a given COSE Algorithm
-func HasherFromCOSEAlg(coseAlg COSEAlgorithmIdentifier) func() hash.Hash {
-	for _, details := range signatureAlgorithmDetails {
-		if details.coseAlg == coseAlg {
-			return details.hasher
-		}
-	}
-	// default to SHA256?  Why not.
-	return crypto.SHA256.New
 }
